1. Changed to type long int:   N and its relatives (neq, Nlocal, Nglobal, etc.);   bandwidth parameters (ml, mu, smu, etc.);   internal indices (i, j, ngroups, etc.) that can have size ~ N;   pivot array, renamed lpivots;   last return flag optional output lsflag; with the following exceptions, where these are of type int:   N, ml, mu, etc. in the C user's call to CVLapackDense and CVLapackBand;   N etc. in the Fortran user's call to FCVLAPACKDENSE and FCVLAPACKBAND   (to be declared as Fortran equivalent of int). (But N, ml, mu arguments in LAPACK user Jacobian fn are still long int.) 2. Interface functions to Blas/Lapack routines declare local int copies of N etc. (intn etc.), and &intn is passed to the BLAS and LAPACK functions in place of &n, etc. 3. Direct linear solver memory block has arrays pivots (int) and lpivots (long int).  File sundials_direct.c has functions NewIntArray and NewLintArray, to create pivots and lpivots, respectively.  LAPACK interfaces use pivots, all other direct solvers use lpivots.

+++ examples/cvode/parallel/cvDiurnal_kry_bbd_p.c
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.2 $
- * $Date: 2008-12-29 22:22:57 $
+ * $Revision: 1.3 $
+ * $Date: 2010-12-01 22:52:20 $
  * -----------------------------------------------------------------
  * Programmer(s): S. D. Cohen, A. C. Hindmarsh, M. R. Wittman, and
  *                Radu Serban  @ LLNL
@@ -167,7 +167,7 @@ int main(int argc, char *argv[])
   realtype abstol, reltol, t, tout;
   N_Vector u;
   int iout, my_pe, npes, flag, jpre;
-  int neq, local_N, mudq, mldq, mukeep, mlkeep;
+  long int neq, local_N, mudq, mldq, mukeep, mlkeep;
   MPI_Comm comm;
 
   data = NULL;
