+++ src/ida/fcmix/fidabbd.c
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.5 $
- * $Date: 2007-04-30 19:29:00 $
+ * $Revision: 1.6 $
+ * $Date: 2010-12-01 22:37:20 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Aaron Collier @ LLNL
  * -----------------------------------------------------------------
@@ -40,11 +40,11 @@
 extern "C" {
 #endif
 
-  extern void FIDA_GLOCFN(int*, 
+  extern void FIDA_GLOCFN(long int*, 
                           realtype*, realtype*, realtype*, realtype*, 
                           long int*, realtype*,
                           int*);
-  extern void FIDA_COMMFN(int*, 
+  extern void FIDA_COMMFN(long int*, 
                           realtype*, realtype*, realtype*, 
                           long int*, realtype*,
                           int*);
@@ -55,8 +55,8 @@ extern "C" {
 
 /*************************************************/
 
-void FIDA_BBDINIT(int *Nloc, int *mudq, int *mldq,
-		  int *mu, int *ml, realtype *dqrely, int *ier)
+void FIDA_BBDINIT(long int *Nloc, long int *mudq, long int *mldq,
+		  long int *mu, long int *ml, realtype *dqrely, int *ier)
 {
   *ier = IDABBDPrecInit(IDA_idamem, *Nloc, *mudq, *mldq, *mu, *ml,
                         *dqrely, (IDABBDLocalFn) FIDAgloc, (IDABBDCommFn) FIDAcfn);
@@ -66,7 +66,7 @@ void FIDA_BBDINIT(int *Nloc, int *mudq, int *mldq,
 
 /*************************************************/
 
-void FIDA_BBDREINIT(int *Nloc, int *mudq, int *mldq,
+void FIDA_BBDREINIT(long int *Nloc, long int *mudq, long int *mldq,
 		    realtype *dqrely, int *ier)
 {
   *ier = 0;
@@ -78,7 +78,7 @@ void FIDA_BBDREINIT(int *Nloc, int *mudq, int *mldq,
 
 /*************************************************/
 
-int FIDAgloc(int Nloc, realtype t, N_Vector yy, N_Vector yp,
+int FIDAgloc(long int Nloc, realtype t, N_Vector yy, N_Vector yp,
 	     N_Vector gval, void *user_data)
 {
   realtype *yy_data, *yp_data, *gval_data;
@@ -110,7 +110,7 @@ int FIDAgloc(int Nloc, realtype t, N_Vector yy, N_Vector yp,
 
 /*************************************************/
 
-int FIDAcfn(int Nloc, realtype t, N_Vector yy, N_Vector yp,
+int FIDAcfn(long int Nloc, realtype t, N_Vector yy, N_Vector yp,
 	    void *user_data)
 {
   realtype *yy_data, *yp_data;
